requirements.txt
cffi==1.17.1
cryptography==45.0.4
et_xmlfile==2.0.0
msoffcrypto-tool==5.4.2
olefile==0.47
openpyxl==3.1.5
pycparser==2.22
PyYAML==6.0.2
rarfile==4.1
